requirements.txt
APScheduler==3.6.3
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
decorator==4.4.2
Fiona==1.8.19
geopandas==0.9.0
haversine==2.3.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
munch==2.5.0
networkx==2.5.1
numpy==1.20.2
osmnx==1.1.0
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
Rtree==0.9.7
scikit-learn==0.24.2
scipy==1.6.3
Shapely==1.7.1
six==1.16.0
staticmap==0.5.5
threadpoolctl==2.1.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
